+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## All the commits and changes that I've made during the entire process
 
 07.04.2022: Added documentationl, insertion of the tree and input numbers from .txt file  
-11.04.2022: Fixed typos mistakes, added first three traversals: inOrder, preOrder, postOrder   
-
-tbd..
+11.04.2022: Fixed typos mistakes, added first three traversals: inOrder, preOrder, postOrder 
+14.04.2022: Fixed some bugs (traversals wouldn't work because the first node is always NULL and I didn't check if node was different than NULL), added the last traversal, levelOrder traversal, changed MAXNODES constant to 20 as I'm going to add only 20 random nodes between 0 and 100, using srand() function     
+**Project done!**